+++ lib/src/create_error.dart
@@ -2,7 +2,11 @@ import 'dart:io';
 
 class CreateError implements HttpException {
   @override
+
+  /// Message to send to client.
   final String message;
+
+  /// HTTP status code (defaults to 400).
   final int errorCode;
 
   CreateError({